Add files via upload
This is a program that will display several lists using a function. Have the user use the list's index position to choose a book. Then a loop will run for the user to select more books. There is a seprate display function for the books the user specifically selected. The last function explains the total of all selected books including a 5% tax.